+++ src/chrono_interop.rs
@@ -41,8 +41,8 @@ use crate::absolute::{AbsoluteTime, CalendarTime};
 /// use chrono::NaiveDateTime;
 /// use chrono::Datelike;
 ///
-/// let mut t = AbsoluteTime::new(100, 12, 30, 25, 340_000_000).unwrap();
-/// t.set_year(Some(2025)).unwrap();
+/// let t = AbsoluteTime::new(100, 12, 30, 25, 340_000_000).unwrap()
+///     .with_year(Some(2025)).unwrap();
 /// let dt: NaiveDateTime = t.into();
 /// assert_eq!(dt.year(), 2025);
 /// assert_eq!(dt.ordinal(), 100);
@@ -131,8 +131,10 @@ mod tests {
 
     #[test]
     fn absolute_to_chrono() {
-        let mut t = AbsoluteTime::new(100, 12, 30, 25, 340_000_000).unwrap();
-        t.set_year(Some(2025)).unwrap();
+        let t = AbsoluteTime::new(100, 12, 30, 25, 340_000_000)
+            .unwrap()
+            .with_year(Some(2025))
+            .unwrap();
         let dt: NaiveDateTime = t.into();
         assert_eq!(dt.year(), 2025);
         assert_eq!(dt.ordinal(), 100);